moving loggers into their own package
