+++ src/main/java/de/idrinth/waraddonclient/CliMain.java
@@ -4,9 +4,9 @@ import de.idrinth.waraddonclient.cli.AliasOption;
 import de.idrinth.waraddonclient.cli.AliasParser;
 import de.idrinth.waraddonclient.service.Config;
 import de.idrinth.waraddonclient.model.CmdAddonList;
-import de.idrinth.waraddonclient.service.CliLogger;
+import de.idrinth.waraddonclient.service.logger.CliLogger;
 import de.idrinth.waraddonclient.service.FileSystem;
-import de.idrinth.waraddonclient.service.MultiLogger;
+import de.idrinth.waraddonclient.service.logger.MultiLogger;
 import de.idrinth.waraddonclient.service.Request;
 import de.idrinth.waraddonclient.service.XmlParser;
 import javax.xml.parsers.ParserConfigurationException;
