+++ src/main/java/de/idrinth/waraddonclient/GuiMain.java
@@ -8,8 +8,8 @@ import de.idrinth.waraddonclient.model.GuiAddonList;
 import de.idrinth.waraddonclient.service.Backup;
 import de.idrinth.waraddonclient.service.FileSystem;
 import de.idrinth.waraddonclient.service.FileWatcher;
-import de.idrinth.waraddonclient.service.GuiLogger;
-import de.idrinth.waraddonclient.service.MultiLogger;
+import de.idrinth.waraddonclient.service.logger.GuiLogger;
+import de.idrinth.waraddonclient.service.logger.MultiLogger;
 import de.idrinth.waraddonclient.service.Request;
 import de.idrinth.waraddonclient.service.Restarter;
 import de.idrinth.waraddonclient.service.Shedule;
