+++ src/main/java/de/idrinth/waraddonclient/Utils.java
@@ -1,6 +1,6 @@
 package de.idrinth.waraddonclient;
 
-import de.idrinth.waraddonclient.service.BaseLogger;
+import de.idrinth.waraddonclient.service.logger.BaseLogger;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
