+++ src/main/java/de/idrinth/waraddonclient/model/GuiAddonList.java
@@ -2,7 +2,7 @@ package de.idrinth.waraddonclient.model;
 
 import de.idrinth.waraddonclient.service.Config;
 import de.idrinth.waraddonclient.Utils;
-import de.idrinth.waraddonclient.service.BaseLogger;
+import de.idrinth.waraddonclient.service.logger.BaseLogger;
 import de.idrinth.waraddonclient.service.Request;
 import de.idrinth.waraddonclient.service.XmlParser;
 import java.awt.event.ActionListener;
