+++ src/main/java/de/idrinth/waraddonclient/service/FileWatcher.java
@@ -1,5 +1,6 @@
 package de.idrinth.waraddonclient.service;
 
+import de.idrinth.waraddonclient.service.logger.BaseLogger;
 import de.idrinth.waraddonclient.model.GuiAddonList;
 import java.io.File;
 import java.io.IOException;
