+++ src/main/java/de/idrinth/waraddonclient/service/Request.java
@@ -1,5 +1,6 @@
 package de.idrinth.waraddonclient.service;
 
+import de.idrinth.waraddonclient.service.logger.BaseLogger;
 import de.idrinth.waraddonclient.Utils;
 import java.io.File;
 import java.io.IOException;
