+++ src/main/java/de/idrinth/waraddonclient/service/logger/BaseLogger.java
@@ -1,11 +1,11 @@
-package de.idrinth.waraddonclient.service;
+package de.idrinth.waraddonclient.service.logger;
 
 public abstract class BaseLogger {
-    protected static final String info = "INFO";
+    protected static final String LEVEL_INFO = "INFO";
 
-    protected static final String error = "ERROR";
+    protected static final String LEVEL_ERROR = "ERROR";
 
-    protected static final String warning = "WARN";
+    protected static final String LEVEL_WARNING = "WARN";
 
     protected abstract void log(String message, String severity);
 
@@ -14,7 +14,7 @@ public abstract class BaseLogger {
     }
 
     public void info(String message) {
-        log(message, info);
+        log(message, LEVEL_INFO);
     }
 
     public void warn(Throwable message) {
@@ -22,7 +22,7 @@ public abstract class BaseLogger {
     }
 
     public void warn(String message) {
-        log(message, warning);
+        log(message, LEVEL_WARNING);
     }
 
     public void error(Throwable message) {
@@ -30,6 +30,6 @@ public abstract class BaseLogger {
     }
 
     public void error(String message) {
-        log(message, error);
+        log(message, LEVEL_ERROR);
     }
 }
