+++ src/main/java/de/idrinth/waraddonclient/service/logger/GuiLogger.java
@@ -1,4 +1,4 @@
-package de.idrinth.waraddonclient.service;
+package de.idrinth.waraddonclient.service.logger;
 
 import javax.swing.JOptionPane;
 
@@ -6,7 +6,7 @@ public class GuiLogger extends BaseLogger {
 
     @Override
     protected void log(String message, String severity) {
-        if (!severity.equals(error)) {
+        if (!severity.equals(LEVEL_ERROR)) {
             return;
         }
         JOptionPane.showMessageDialog(null, message);
